Exercise 2
Contains the optimisation so far for part 2 of the project. 

Done: Moved "walls" and "goals" to shared arrays
Still needs to be done: Set MAX_COL and MAX_ROW before init.

REMARK: I couldn't figure out how to read the MAX_COL and MAX_ROW from the file before going through it since it is the server that reads the file and not our code. 

So right now it defaults to 70x70 and then save the MAX_COL and MAX_ROW after first read. 

But that means some states still save 70x70 arrays for the boxes - but only loops through the actual max values.